add explicit cast to double

+++ evolve/src/main/java/org/dishevelled/evolve/select/RankBasedSelection.java
@@ -99,7 +99,7 @@ public final class RankBasedSelection<I>
         int intermediateSize = intermediate.size();
         for (I i : intermediate.keySet())
         {
-            intermediate.put(i, Double.valueOf(rank / intermediateSize));
+            intermediate.put(i, Double.valueOf((double) rank / (double) intermediateSize));
         }
 
         // fitness proportional selection on intermediate map
